Django-signals_orm-0x04/requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
dj-rest-auth==7.0.1
idna==3.10
mysqlclient==2.2.7
nose==1.3.7
parameterized==0.9.0
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0

